[JXD-231] Вынес класс интерфейса как отдельный компонент
