+++ esphome/components/display_menu_base/__init__.py
@@ -26,6 +26,8 @@ from esphome.const import (
 
 CODEOWNERS = ["@numo68"]
 
+AUTO_LOAD = ["display_menu_render_base"]
+
 display_menu_base_ns = cg.esphome_ns.namespace("display_menu_base")
