+++ esphome/components/display_menu_base/display_menu_base.h
@@ -1,12 +1,10 @@
 #pragma once
 
 #include "esphome/core/component.h"
-
+#include "esphome/components/display_menu_render_base/display_menu_render_base.h"
 #include "menu_item.h"
 
 #include <forward_list>
-
-#include "esphome/components/groups/entity_types.h"
 
 namespace esphome {
 namespace display_menu_base {
@@ -19,22 +17,7 @@ enum MenuMode {
 class MenuItem;
 class MenuItemMenu;
 
-// Render interface for menu with groups
-class MenuRenderInterface {
- public:
-  MenuRenderInterface(groups::EntityType type, groups::EntitySubtype subtype = groups::EntitySubtype::NONE)
-      : type_(type), subtype_(subtype){};
-
-  groups::EntityType type() { return type_; }
-  groups::EntitySubtype subtype() { return subtype_; }
-
-  // Add needed items in menu with entity_info
-  virtual size_t render_entity(MenuItemMenu *menu, const groups::EntityInfo &info) = 0;
-
- protected:
-  groups::EntityType type_;
-  groups::EntitySubtype subtype_;
-};
+using namespace display_menu_render_base;
 
 /** Class to display a hierarchical menu.
  *
