+++ esphome/components/display_menu_render_dallas/display_menu_render_dallas.h
@@ -1,13 +1,13 @@
 #pragma once
 
-#include "esphome/components/display_menu_base/display_menu_base.h"
-#include "esphome/components/groups/entity_types.h"
+#include "esphome/components/display_menu_render_base/display_menu_render_base.h"
 #include "esphome/components/dallas_temp/dallas_temp.h"
 
 namespace esphome {
 namespace display_menu_render_dallas {
 
 using namespace display_menu_base;
+using namespace display_menu_render_base;
 
 class DallasTempMenuRender : public MenuRenderInterface {
  public:
